% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map State Motor Vehicle Fatalities.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numerical code for US state.}

\item{year}{value of year}
}
\description{
This function takes a state number and a year, and draws
a state outline with dots to represent the location of
motor vehicle fatalities for that year.
}
\details{
\code{fars_map_state} maps state motor vehicle fatalities #' given a year and state id number.
}
\examples{
  \dontrun{
    fars_map_state(48, 2013)
    fars_map_state(48, 2014)
    fars_map_state(6, 2013)
  }

}
