% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of file that contains FARS data}
}
\value{
This function returns a data frame containing the
		FARS data. If an incorrect filename is entered the
	function will stop and print 'does not exist'
	message.
}
\description{
This function reads data from the US National Highway
Traffic Safety Administration's [Fatality Analysis
Reporting System](https://www.nhtsa.gov/Data/Fatality-
Analysis-Reporting-System-(FARS)), given a filename for the
data. It returns a tibble of the data. For this function to
work properly, a filename pointing to an existing file must
be given.
}
\details{
\code{fars_read} reads FARS data in R envirnoment.
}
\examples{
\dontrun{
fars_read(accident_2014.csv.bz2)
}

}
