% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS files for one or more years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of years}
}
\value{
This function returns a list of tibbles, where each #' tibble contains containing the year and month from the
observations in the corresponding year's FARS data.
}
\description{
This function takes a vector of years, pulls the FARS data #' for those years, and then produces a summary data frames. #' The summary data frames shows the number of observations
for each month/year combination for the extracted FARS
data.
}
\details{
\code{fars_read_years} produces a list of data frames of
FARS data, based on input vector of years passed as
arguments.
}
\examples{
\dontrun{
fars_read_years(years = c(2013, 2014, 2015))
fars_read_years(years = 2013)
}

}
