% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Produce a Summary of FARS Files.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years}
}
\value{
This function returns tibble where the first column #' is the month,the second and following columns are the
requested years, and the rows for the year columns are the
number of FARS observations for that month/year
combination.
}
\description{
This function takes a vector of years as input and produces #' a list of data frames corresponding to input given
}
\details{
\code{fars_summarize_years} produces a summary tibble of
FARS years and months for given a vector of years
}
\examples{
\dontrun{
fars_summarize_years(years = c(2013, 2014, 2015))
fars_summarize_years(years = 2013)
}

}
