% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS Filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{value of the year}
}
\value{
This function returns a a FARS filename.
}
\description{
This function takes a year as input and produces a FARS
filename. Enter values as per requirement
}
\details{
\code{make_filename} This will return FARS
	filename based on year passed.
}
\examples{
\dontrun{
		make_filename(2013) 
	}
}
