% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display map of accidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The integer for a state code}

\item{year}{The year of the data for that state}
}
\value{
No value returned, but creates a map with the accidents for that year. An error will be returned if you give an invalid state number, or if there are no accidents for the parameters given.
}
\description{
Creates a state map with the accidents plotted with latitud and longitude.
}
\examples{
fars_map_state(17, 2015)

}
