% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of State}

\item{year}{year of data}
}
\value{
Retruns map of selected state and location of accidents
\cr Issues error "invalid State number" if state number provided does not exist
\cr Issues message if no records are retrieved for the selection
}
\description{
Map of accidents
}
\examples{
fars_map_state(1,2015)
}
