% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show incident locations in a 2D plot}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state numnber}

\item{year}{year number}
}
\value{
map of the state, with incident locations.
}
\description{
Takes state number and year, then read and filter from file, and show in a 2D plot.
}
\details{
year should be four digits; state number must exist in file. if invalid state number, it will stop with error.
}
