% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read the file with the FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to be read}
}
\value{
A dataframe of with the data that was read in, or an error if it is an invalid file name.
}
\description{
Reads in the file name of the FARS data and then stores it in a data frame.
}
\examples{
fars_read(accident_2015.csv.bz2)

}
