% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{file name of data}
}
\value{
Extracts csv data
\cr On error, message "file... does not exist" is issued.
}
\description{
Read CSV
}
\examples{
\donttest{
fars_read("accident_2013.csv.bz2")}
}
