% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string of file name}
}
\value{
data.frame from csv file contents
}
\description{
Takes file name and read to memory
}
\details{
If file does not exist, it will stop with an error message.
}
