% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in the years to be collected and organized into month and year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The year or years you want to check data for.}
}
\value{
A dataframe of with the data for all the years selected, or an error if it is an invalid file name or invalid year.
}
\description{
Creates a data frame with month and year if the file name is correct.
}
\examples{
fars_read_years(2015)
}
