% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Filter by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{calendar year of selection}
}
\value{
Returns records of accidents for the selected year by month.
\cr On error, warning "invalid year" is issued
}
\description{
Filter by year
}
\examples{
fars_read_years(2015)
}
