% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Assemble data from multiple year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{multiple year in a vector}
}
\value{
data frame of month/year
}
\description{
Takes year vector to commpose file names and read the files and put month/year to data frame.
}
\details{
year should be four digits, although not mandatory, files must be exisiting.
}
