% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of the years you want to summarize.}
}
\value{
A data frame with the summary by year, then month. No error returned in this function, as validation is done elsewhere.
}
\description{
Creates a data frame with a summary of the years specified.
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
