% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of accidents by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{calender year of accidents to be retrieved}
}
\value{
summary of accidents for the year by month
}
\description{
Summary of accidents by year
}
\examples{
\donttest{
fars_summarize_years(2015)}
}
