% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of year values.}
}
\value{
This function returns a data frame containing the number of observations
for each month under the accident data files of each of the corresponding years.
}
\description{
This function takes in an argument "years" containing a list of year values, and calls
the function fars_read_years to generate a list of data frames containing the MONTH
and year values of the accident files corresponding to these years. It then calls the
bind_rows function from dplyr package to combine these data frames into a single data
frame, then returns the count for each of the twelve months of each of the years
specified under "years".
}
\examples{
dat_sum <- fars_summarize_years (c(2013, 2014))

}
