% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize by year, spread to months}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{multiple year in a vector}
}
\value{
summary data frame by year, with each month on columns
}
\description{
Takes year vector then read the files of these years and count by year to the level of month.
}
\details{
year should be four digits, although not mandatory, files must be exisiting.
}
