% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes a file name for the years requested}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the data you want to analyze as an integer.}
}
\value{
A string of the correct file name for that year. No error will be returned as long as you send an integer, or string that can be converted to an integer as the parameter.
}
\description{
The main function to create the file name for the year.
}
\examples{
make_filename(2015)

}
