% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a file name following the prescribed format}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year data was collected}
}
\value{
File name using  the prescribed format
\cr concatenates accident_+year+.csv.bz2
}
\description{
Creates a file name following the prescribed format
}
\examples{
make_filename(2015)
}
