% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A value to be incorporated into the file name to be created.}
}
\value{
This function returns a character string of the form "accident_"year".csv.bz2".
}
\description{
This function takes in a argument "year", coerces it into an integer,
and returns a character string of the form "accident_"year".csv.bz2".
}
\examples{
make_filename (2015)

}
