% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year number}
}
\value{
string of file name
}
\description{
Takes year to compose file name
}
\details{
year should be four digits, although not mandatory
}
